+++ src/main/java/mil/nga/geopackage/extension/nga/index/FeatureTableIndex.java
@@ -59,8 +59,24 @@ public class FeatureTableIndex extends FeatureTableCoreIndex {
 	 *            feature dao
 	 */
 	public FeatureTableIndex(GeoPackage geoPackage, FeatureDao featureDao) {
+		this(geoPackage, featureDao, false);
+	}
+
+	/**
+	 * Constructor
+	 * 
+	 * @param geoPackage
+	 *            GeoPackage
+	 * @param featureDao
+	 *            feature dao
+	 * @param geodesic
+	 *            index using geodesic bounds
+	 * @since 6.6.5
+	 */
+	public FeatureTableIndex(GeoPackage geoPackage, FeatureDao featureDao,
+			boolean geodesic) {
 		super(geoPackage, featureDao.getTableName(),
-				featureDao.getGeometryColumnName());
+				featureDao.getGeometryColumnName(), geodesic);
 		this.featureDao = featureDao;
 	}