+++ src/test/java/mil/nga/geopackage/extension/rtree/RTreeIndexExtensionImportTest.java
@@ -22,14 +22,27 @@ public class RTreeIndexExtensionImportTest extends ImportGeoPackageTestCase {
 
 	/**
 	 * Test RTree
-	 * 
+	 *
 	 * @throws SQLException
 	 *             upon error
 	 */
 	@Test
 	public void testRTree() throws SQLException {
 
-		RTreeIndexExtensionUtils.testRTree(geoPackage);
+		RTreeIndexExtensionUtils.testRTree(geoPackage, false);
+
+	}
+
+	/**
+	 * Test RTree with geodesic
+	 *
+	 * @throws SQLException
+	 *             upon error
+	 */
+	@Test
+	public void testRTreeGeodesic() throws SQLException {
+
+		RTreeIndexExtensionUtils.testRTree(geoPackage, true);
 
 	}