+++ src/test/java/mil/nga/geopackage/tiles/features/FeatureTileGeneratorTest.java
@@ -42,7 +42,7 @@ public class FeatureTileGeneratorTest extends CreateGeoPackageTestCase {
 	 */
 	@Test
 	public void testTileGenerator() throws IOException, SQLException {
-		testTileGenerator(false, false, false);
+		testTileGenerator(false, false, false, false);
 	}
 
 	/**
@@ -53,7 +53,7 @@ public class FeatureTileGeneratorTest extends CreateGeoPackageTestCase {
 	 */
 	@Test
 	public void testTileGeneratorWithIndex() throws IOException, SQLException {
-		testTileGenerator(true, false, false);
+		testTileGenerator(true, false, false, false);
 	}
 
 	/**
@@ -64,7 +64,7 @@ public class FeatureTileGeneratorTest extends CreateGeoPackageTestCase {
 	 */
 	@Test
 	public void testTileGeneratorWithIcon() throws IOException, SQLException {
-		testTileGenerator(false, true, false);
+		testTileGenerator(false, true, false, false);
 	}
 
 	/**
@@ -76,7 +76,19 @@ public class FeatureTileGeneratorTest extends CreateGeoPackageTestCase {
 	@Test
 	public void testTileGeneratorWithMaxFeatures()
 			throws IOException, SQLException {
-		testTileGenerator(false, false, true);
+		testTileGenerator(false, false, true, false);
+	}
+
+	/**
+	 * Test tile generator
+	 *
+	 * @throws java.io.IOException
+	 * @throws java.sql.SQLException
+	 */
+	@Test
+	public void testTileGeneratorWithGeodesic()
+			throws IOException, SQLException {
+		testTileGenerator(false, false, false, true);
 	}
 
 	/**
@@ -88,7 +100,19 @@ public class FeatureTileGeneratorTest extends CreateGeoPackageTestCase {
 	@Test
 	public void testTileGeneratorWithIndexAndIcon()
 			throws IOException, SQLException {
-		testTileGenerator(true, true, false);
+		testTileGenerator(true, true, false, false);
+	}
+
+	/**
+	 * Test tile generator
+	 *
+	 * @throws java.io.IOException
+	 * @throws java.sql.SQLException
+	 */
+	@Test
+	public void testTileGeneratorWithIndexAndIconAndGeodesic()
+			throws IOException, SQLException {
+		testTileGenerator(true, true, false, true);
 	}
 
 	/**
@@ -100,21 +124,41 @@ public class FeatureTileGeneratorTest extends CreateGeoPackageTestCase {
 	@Test
 	public void testTileGeneratorWithIndexAndIconAndMaxFeatures()
 			throws IOException, SQLException {
-		testTileGenerator(true, true, true);
+		testTileGenerator(true, true, true, false);
+	}
+
+	/**
+	 * Test tile generator
+	 *
+	 * @throws java.io.IOException
+	 * @throws java.sql.SQLException
+	 */
+	@Test
+	public void testTileGeneratorWithIndexAndIconAndMaxFeaturesAndGeodesic()
+			throws IOException, SQLException {
+		testTileGenerator(true, true, true, true);
 	}
 
 	/**
 	 * Test tile generator
 	 *
 	 * @param index
+	 *            index features
 	 * @param useIcon
+	 *            true to use an icon instead of the default point
 	 * @param maxFeatures
+	 *            set max features
+	 * @param geodesic
+	 *            draw geometries using geodesic lines
 	 *
 	 * @throws java.io.IOException
+	 *             upon error
 	 * @throws java.sql.SQLException
+	 *             upon error
 	 */
 	public void testTileGenerator(boolean index, boolean useIcon,
-			boolean maxFeatures) throws IOException, SQLException {
+			boolean maxFeatures, boolean geodesic)
+			throws IOException, SQLException {
 
 		int minZoom = 0;
 		int maxZoom = 4;
@@ -124,11 +168,11 @@ public class FeatureTileGeneratorTest extends CreateGeoPackageTestCase {
 		int num = FeatureTileUtils.insertFeatures(geoPackage, featureDao);
 
 		FeatureTiles featureTiles = FeatureTileUtils
-				.createFeatureTiles(geoPackage, featureDao, useIcon);
+				.createFeatureTiles(geoPackage, featureDao, useIcon, geodesic);
 
 		if (index) {
 			FeatureIndexManager indexManager = new FeatureIndexManager(
-					geoPackage, featureDao);
+					geoPackage, featureDao, geodesic);
 			featureTiles.setIndexManager(indexManager);
 			indexManager.setIndexLocation(FeatureIndexType.GEOPACKAGE);
 			int indexed = indexManager.index();
@@ -194,8 +238,9 @@ public class FeatureTileGeneratorTest extends CreateGeoPackageTestCase {
 
 		TestCase.assertEquals(expectedTiles, tiles);
 
-		// TileWriter.writeTiles(geoPackage, "gen_feature_tiles", new File(
-		// "/Users/osbornb/Documents/generator/tiles"), null, null, true);
+		// TileWriter.writeTiles(geoPackage, "gen_feature_tiles",
+		// new File("/Users/osbornb/Documents/generator/tiles"), null,
+		// null, null, TileFormatType.XYZ, true);
 
 	}