+++ src/test/java/mil/nga/geopackage/tiles/features/FeatureTilesTest.java
@@ -33,7 +33,17 @@ public class FeatureTilesTest extends CreateGeoPackageTestCase {
 	 */
 	@Test
 	public void testFeatureTiles() throws SQLException {
-		testFeatureTiles(false);
+		testFeatureTiles(false, false);
+	}
+
+	/**
+	 * Test feature tiles
+	 *
+	 * @throws java.sql.SQLException
+	 */
+	@Test
+	public void testFeatureTilesWithGeodesic() throws SQLException {
+		testFeatureTiles(false, true);
 	}
 
 	/**
@@ -43,29 +53,43 @@ public class FeatureTilesTest extends CreateGeoPackageTestCase {
 	 */
 	@Test
 	public void testFeatureTilesWithIcon() throws SQLException {
-		testFeatureTiles(true);
+		testFeatureTiles(true, false);
+	}
+
+	/**
+	 * Test feature tiles
+	 *
+	 * @throws java.sql.SQLException
+	 */
+	@Test
+	public void testFeatureTilesWithIconGeodesic() throws SQLException {
+		testFeatureTiles(true, true);
 	}
 
 	/**
 	 * Test feature tiles
 	 *
 	 * @param useIcon
+	 *            true to use an icon instead of the default point
+	 * @param geodesic
+	 *            draw geometries using geodesic lines
 	 *
 	 * @throws java.sql.SQLException
 	 */
-	public void testFeatureTiles(boolean useIcon) throws SQLException {
+	public void testFeatureTiles(boolean useIcon, boolean geodesic)
+			throws SQLException {
 
 		FeatureDao featureDao = FeatureTileUtils.createFeatureDao(geoPackage);
 
 		int num = FeatureTileUtils.insertFeatures(geoPackage, featureDao);
 
 		FeatureTiles featureTiles = FeatureTileUtils
-				.createFeatureTiles(geoPackage, featureDao, useIcon);
+				.createFeatureTiles(geoPackage, featureDao, useIcon, geodesic);
 
 		try {
 
 			FeatureIndexManager indexManager = new FeatureIndexManager(
-					geoPackage, featureDao);
+					geoPackage, featureDao, geodesic);
 			featureTiles.setIndexManager(indexManager);
 
 			indexManager.setIndexLocation(FeatureIndexType.GEOPACKAGE);