+++ CHANGELOG.md
@@ -10,6 +10,8 @@ Adheres to [Semantic Versioning](http://semver.org/).
 * sqlite-jdbc version 3.45.1.0
 * Get abstract User DAO by table name
 * Set User Custom DAO contents
+* Feature Tiles geodesic draw support
+* Feature Index Manager, RTree Index, Feature Table Index, and Manual query geodesic support
 
 ## [6.6.4](https://github.com/ngageoint/geopackage-java/releases/tag/6.6.4) (11-29-2023)